requirements.txt
natsort==8.4.0
numpy==2.1.1
PDFNetPython3==9.5.0.post2
pillow==10.4.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dotenv==1.0.1
